Add statically compiled gRPC code, add tests for it
This commit adds another, static way of using this library.

Instead of compiling the proto files on the fly when the library is
initializated, this second static approach uses pre-compiled files.

The main benefit, beside probably the initialization being slightly
faster, is that the auto-completion is better due to also having
TypeScript declaration files in parallel to the compiled JavaScript code
itself.